+++ js/js.translator/src/org/jetbrains/kotlin/js/translate/declaration/ClassTranslator.kt
@@ -90,7 +90,7 @@ class ClassTranslator private constructor(
         val nonConstructorContext = context.innerWithUsageTracker(scope, descriptor)
         val delegationTranslator = DelegationTranslator(classDeclaration, nonConstructorContext)
         translatePropertiesAsConstructorParameters(nonConstructorContext, properties)
-        val bodyVisitor = DeclarationBodyVisitor(properties, staticProperties)
+        val bodyVisitor = DeclarationBodyVisitor(properties, staticProperties, scope)
         bodyVisitor.traverseContainer(classDeclaration, nonConstructorContext)
         delegationTranslator.generateDelegated(properties)
 
@@ -361,9 +361,9 @@ class ClassTranslator private constructor(
     }
 
     private fun generateBridgesToTraitImpl(properties: MutableList<JsPropertyInitializer>) {
-        for (entry in CodegenUtil.getNonPrivateTraitMethods(descriptor).entries) {
-            if (!areNamesEqual(entry.key, entry.value)) {
-                properties += generateDelegateCall(entry.value, entry.key, JsLiteral.THIS, context())
+        for ((key, value) in CodegenUtil.getNonPrivateTraitMethods(descriptor)) {
+            if (!areNamesEqual(key, value)) {
+                properties += generateDelegateCall(value, key, JsLiteral.THIS, context())
             }
         }
     }