+++ js/js.translator/testData/inline/cases/anonymousFunction.kt
@@ -1,0 +1,12 @@
+package foo
+
+// CHECK_CONTAINS_NO_CALLS: test
+
+inline fun <T> block(f: () -> T): T = f()
+
+fun test() = block(fun(): Int { return 23 })
+
+fun box(): String {
+    assertEquals(23, test())
+    return "OK"
+}