+++ ChangeLog.md
@@ -9,6 +9,10 @@
 
 ### Compiler
 
+### JS
+
+- Improve performance of maps and sets
+
 #### Analysis & diagnostics
 
 - Combination of 'open' and 'override' is no longer a warning
@@ -30,7 +34,7 @@
 - [`KT-11499`](https://youtrack.jetbrains.com/issue/KT-11499) Compiler crashes with "Incompatible stack heights"
 - [`KT-11943`](https://youtrack.jetbrains.com/issue/KT-11943) CompilationException with extension property of KClass
 - [`KT-12125`](https://youtrack.jetbrains.com/issue/KT-12125) Wrong increment/decrement on Byte/Char/Short.MAX_VALUE/MIN_VALUE
-- [`KT-12192`](https://youtrack.jetbrains.com/issue/KT-12192) Exhaustiveness check isn't generated for when expression returing Unit
+- [`KT-12192`](https://youtrack.jetbrains.com/issue/KT-12192) Exhaustiveness check isn't generated for when expression returning Unit
 
 #### Performance
 
@@ -213,6 +217,7 @@
 ###### New features
 
 - Added "Decompile" button to Kotlin bytecode toolwindow
+- [`KT-11657`](https://youtrack.jetbrains.com/issue/KT-11657) Allow viewing decompiled Java source code for Kotlin-compiled classes
 - [`KT-11704`](https://youtrack.jetbrains.com/issue/KT-11704) Support file path references inside of Kotlin string literals
 
 ###### Issues fixed: